backend support for overloaded access keys
i wanted to ensure that salts are always unique so i made them the primary key in the table. this makes it costly to ensure that access keys are unique so as a side effect, access keys can be overloaded. when multiple items are returned for the same access key, the client needs some (cheap) way to figure out which salt to use to generate the auth token (since argon2 hashing is expensive).
this is a rare edge case given the 7700^3 [currently] possible access key space but can still happen...
the select id is easy to compute being just a sah256 hash of the passphrase (also not guaranteed to be unique) and salt (guranteed to be unique).

this can leaks the existance of multiple secrets for the same access key but it's not clear to me what advantage an attacker has with this knowledge.